Distinguish between inline data and values clause
SPARQL 1.1 http://www.w3.org/TR/sparql11-query/#grammar  distinguishes between values clauses [21] and inline data blocks [61], with the difference that the ValuesClasue can also be used at end of the query or subselect.
This fix adds the distinction to the jison grammar, adds a simple test case for values clause at of the query and fixes two test cases.