Modify Generator to bypass entity IRI wrapping
I want to be able to create a SPARQL query without having an expression represented as an AST. eg:
```javascript
{ // ...
  where: {
    { type: 'filter',
      expression: 'regex(?var, "^match")',
    }
  }
}
```

Yields:
```sparql
WHERE {
  # ...
  FILTER(<regex(?var,"^match")>)
}
```

In fact, it would never make sense to have any filter have just an IRI ref. Instead of making the code check for this or adding much complication, why not just allow the user to bypass this?

This seems like a safe way to extend that functionality. Prefixing my string with a space could indicate I don't want entity conversion/handling. Aside from uses forcing a string into the `expression` key, none of the other methods would ever pass a `value` having a space at the 0th character, so this addition is compatible with all other features. This is more of a patch than an API change.