Issue 99 - Fix blank node triple pattern with property paths
When blank node triple patterns of the form `[ :p "v" ] .` contained
property paths, the parsed throwed an exception because it didn't expect
it. This happened because the parser entered the ObjectList rule that
doesn't handle the property path case. So instead of having PropertyListPathNotEmptyTail
calling the ObjectList rule, I added a ObjectListPath rule that is
called instead and that calls GraphNodePath, following the same pattern
of -Path suffixed rules in this parser to make sure that property paths
arent allowed in other queries such as construct queries.

To the unit tests I've added the queries posted in the issue

https://github.com/RubenVerborgh/SPARQL.js/issues/99

Reference: https://www.w3.org/TR/2013/REC-sparql11-query-20130321/#QSynBlankNodes